优化登录代码
